+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Description: Robonaut 2
 - Description: Valkyrie
 - Loader: MuJoCo
 - Unit tests to check that MuJoCo loads all MJCF descriptions successfully